samples/vulnerable-app-2/requirements.txt
Flask==2.3.0
cryptography==41.0.0
pycryptodome==3.18.0
requests==2.31.0
PyJWT==2.8.0
ecdsa==0.18.0
